Move annotatedStringResource
